api/requirements.txt
Faker==0.0.4
Flask==0.10.1
Flask-Assets==0.10
Flask-Script==2.0.5
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.8
Scrapy==1.0.1
Twisted==15.2.1
Werkzeug==0.10.4
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
cffi==1.1.2
cryptography==0.9.3
cssselect==0.9.1
enum34==1.0.4
gevent==1.0.2
greenlet==0.4.7
grequests==0.2.0
idna==2.0
ipaddress==1.0.14
itsdangerous==0.24
kombu==3.0.26
lxml==3.4.4
pyOpenSSL==0.15.1
pyasn1==0.1.8
pycparser==2.14
python-dateutil==2.4.2
pytz==2015.4
queuelib==1.2.2
requests==2.7.0
simplejson==3.8.0
six==1.9.0
w3lib==1.12.0
webassets==0.10.1
wsgiref==0.1.2
zope.interface==4.1.2
